Expand the description of Application support (#8619)
* Expand the description of Application support

I was not aware that the applications can also be auto-discovered. I discovered it by chance, and when I mentioned it in discord channel, it seemed blatantly obvious to everyone else. Thought editing the docs might help future users.

* Update Applications.md

Changed a lot of instructions to make it more structured.

* Add files via upload

Addedimages to help with application extensions how-to.

* Update Applications.md

Added the file locations into master file.